added test to multiply and removed the testmatrix funcs
